% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{alpha_igr}
\alias{alpha_igr}
\title{Alpha param. used in ingestion rate function.
NB: this function works under the assumption that we consider 10 cm water depth as
to work with area instead of volume.
constants from kiorboe2014shifts}
\usage{
alpha_igr(w)
}
\arguments{
\item{w}{\link{g}}
}
\value{
alpha value \link{gr (food source) m-2} , m2 considering 10 cm water depth as
}
\description{
Alpha param. used in ingestion rate function.
NB: this function works under the assumption that we consider 10 cm water depth as
to work with area instead of volume.
constants from kiorboe2014shifts
}
\examples{
alpha_igr(convertL_to_W(5))
alpha_igr(2)

}
