% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClass.R
\name{convertL_to_W}
\alias{convertL_to_W}
\title{Convert size to weight W}
\usage{
convertL_to_W(L)
}
\arguments{
\item{L}{size \link{cm}}
}
\value{
equivalent weight \link{gr}
}
\description{
Convert size to weight W
}
\examples{
convertL_to_W(L = 2.5)
}
