% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{count_devDays}
\alias{count_devDays}
\title{Count how days is needed to growth fro a customarily initial length until reach the max. L}
\usage{
count_devDays(df_col, min_size, max_size)
}
\arguments{
\item{df_col}{column of a DF with data from simulation of somatic growth for different temperatures. The columns the temperature (as Character)}

\item{min_size}{min. Size to start to count}

\item{max_size}{max. where to stop the count}
}
\value{
development days \link{days}
}
\description{
Count how days is needed to growth fro a customarily initial length until reach the max. L
}
\examples{
sapply(development_df_juvI, function(col){ count_devDays(col, juvI_min, juvI_max) } )
}
