% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_temming.R
\name{egg_dev}
\alias{egg_dev}
\title{Development status Egg}
\usage{
egg_dev(dev_status, temperature)
}
\arguments{
\item{dev_status}{is the current dev. status of an individual or cohort. Valid values (0,1]. When status => 1, then egg hatch and change status to 'Larvae'}

\item{temperature}{temperature of the day \link{°C}}
}
\value{
a list with 'new dev. status' a boolean wheather ready to hatch or nor in order to track history of development
}
\description{
Development status Egg
}
\examples{
egg_dev(0.5, 15)
for (i in 2:length(temperature2015_Q2$temperature) ){
egg_development <- egg_dev(current_dev_status, temperature2015_Q2$temperature[i])
current_dev_status <- egg_development[[1]]
hatch_status_array_Q2 <- append(hatch_status_array_Q2, egg_development[[2]] )
egg_development_array_Q2 <- append(egg_development_array_Q2,current_dev_status)}
}
