% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{eta_J}
\alias{eta_J}
\title{Eta function
This function is used only internally as is nested in 'ST_predation'}
\usage{
eta_J(t_day, power_of = 2, J = 250)
}
\arguments{
\item{J}{}
}
\description{
Eta function
This function is used only internally as is nested in 'ST_predation'
}
