% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{hatch_eggs}
\alias{hatch_eggs}
\title{Shift of biomass from eggs to larvae or 'growth function for eggs'.}
\usage{
hatch_eggs(Te)
}
\arguments{
\item{Te}{\link{°C}}
}
\value{
rate at which eggs are hatching \link{day^-1}
}
\description{
Shift of biomass from eggs to larvae or 'growth function for eggs'.
}
\examples{
hatch_eggs(10)
}
