% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{ingestion_kernel}
\alias{ingestion_kernel}
\title{Predation activity
Ingestion kernel function for predation}
\usage{
ingestion_kernel(I_max, l_pred, l_prey)
}
\arguments{
\item{l_pred}{\link{cm}}

\item{l_prey}{\link{cm}}
}
\value{
probability that size class l_prey will be eaten by predator of size l_pred \link{day^-1}
}
\description{
Predation activity
Ingestion kernel function for predation
}
\examples{
ingestion_kernel
}
