% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{ingestion_rate_b}
\alias{ingestion_rate_b}
\title{Growth in weight for an specific size class 'L' in dependence of resouce availabilty. Applies for sizes between (6 and L_inf)}
\usage{
ingestion_rate_b(temperature, L, P, sex)
}
\arguments{
\item{temperature}{Temperature \link{°C}}

\item{L}{size class \link{cm}}

\item{P}{Plancton (variable state)}
}
\value{
growth rate in weight \link{gr / time }
}
\description{
Growth in weight for an specific size class 'L' in dependence of resouce availabilty. Applies for sizes between (6 and L_inf)
}
\examples{
ingestion_rate(temperature = 15, L = 55, P=2, sex = 'F')
}
