% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClass.R
\name{K_func}
\alias{K_func}
\title{*Deprecated Function to calculate K (growth parameter from vB eqs.)}
\usage{
K_func(temperature)
}
\value{
K value as function of temperature
}
\description{
*Deprecated Function to calculate K (growth parameter from vB eqs.)
}
