% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{K_func_briere}
\alias{K_func_briere}
\title{K is the vB growth factor transfored to flexTPC to add
temperature dependency}
\usage{
K_func_briere(temperature, sex_params)
}
\arguments{
\item{temperature}{\link{°C}}
}
\value{
K value
}
\description{
K is the vB growth factor transfored to flexTPC to add
temperature dependency
}
\examples{
K_func_briere(10, parameters_solv$fem_params)
}
