% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_temming.R
\name{larvae_dev}
\alias{larvae_dev}
\title{Development status Larvae}
\usage{
larvae_dev(dev_status, temperature)
}
\arguments{
\item{dev_status}{is the current dev. status of an individual or cohort. Valid values (0,1]. When status => 1, then Larvae changes status to 'Juvenile'}

\item{temperature}{temperature of the day \link{°C}}
}
\value{
a list with 'new dev. status' a boolean wheather ready to hatch or nor in order to track history of development
}
\description{
Development status Larvae
}
\examples{
larvae_dev(0.5, 15)
}
