% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{lopt}
\alias{lopt}
\title{optimal prey length
source of param values: Ovidio (@tbd correct source)}
\usage{
lopt(l_pred)
}
\arguments{
\item{l_pred}{\link{cm}}
}
\value{
the optimal length of prey for a predator of size l_pred
}
\description{
optimal prey length
source of param values: Ovidio (@tbd correct source)
}
\examples{
lopt(2.75)
}
