% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClass.R
\name{molting_fraction}
\alias{molting_fraction}
\title{Fraction of females that are molting and therefore releasing eggs.}
\usage{
molting_fraction(L, Te, Fem_params)
}
\arguments{
\item{L}{size \link{cm}}

\item{temperature}{Temperature \link{°C}}
}
\value{
a fraction \link{dimensionless}
}
\description{
Fraction of females that are molting and therefore releasing eggs.
}
