% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClass.R
\name{natural_mortality}
\alias{natural_mortality}
\title{Natural mortality}
\usage{
natural_mortality(temperature, L)
}
\arguments{
\item{temperature}{Temperature \link{°C}}

\item{L}{size \link{cm}}
}
\value{
\link{gr / time }
}
\description{
Natural mortality
}
