% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{new_Bpredator}
\alias{new_Bpredator}
\title{Predatior population. Oversimplification of predator population with
cosinus preiodic function presenting peaks around sprint.}
\usage{
new_Bpredator(t)
}
\arguments{
\item{t}{\link{days}}
}
\value{
biomass of predator \link{@tbd}
}
\description{
Predatior population. Oversimplification of predator population with
cosinus preiodic function presenting peaks around sprint.
}
