% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClass.R
\name{new_food}
\alias{new_food}
\title{Production of resources (Plancton) - TEMPORARILY ARTIFICIAL FUNCTION WITH COSINUS; TO BE REPLACED BY A MORE REALISTIC ONE}
\usage{
new_food(t, Te, scale = 2)
}
\arguments{
\item{t}{time}
}
\description{
Production of resources (Plancton) - TEMPORARILY ARTIFICIAL FUNCTION WITH COSINUS; TO BE REPLACED BY A MORE REALISTIC ONE
}
