% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{prep_sol_Lavg}
\alias{prep_sol_Lavg}
\title{Title}
\usage{
prep_sol_Lavg(df_sol, init_year, fin_year)
}
\arguments{
\item{init_year}{31/12 of the previous year you want to see at first e.g. 31/12/2010 to see first 2011}

\item{fin_year}{01/01 of the previous year you want to see at first e.g. 01/01/2016 to see until 2015}
}
\description{
Title
}
\examples{
prep_sol_Lavg(df_sol, "31/12/2010")
}
