% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClass.R
\name{respiration_rate}
\alias{respiration_rate}
\title{Natural mortality}
\usage{
respiration_rate(temperature, L)
}
\arguments{
\item{temperature}{Temperature \link{°C}}

\item{L}{size \link{cm}}
}
\value{
\link{gr / time }
}
\description{
Natural mortality
}
