% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{respiration_rate_b}
\alias{respiration_rate_b}
\title{Respiration rate: is the energy loss.
NB: as the model includes other types of mortality, respiration rate
is kept very small, therefore multiplied by 0.1}
\usage{
respiration_rate_b(temperature, L, sex_params)
}
\arguments{
\item{temperature}{Temperature \link{°C}}

\item{L}{size \link{cm}}
}
\value{
\link{g day^-1 }
}
\description{
Respiration rate: is the energy loss.
NB: as the model includes other types of mortality, respiration rate
is kept very small, therefore multiplied by 0.1
}
