% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{shift_next_sizeClass}
\alias{shift_next_sizeClass}
\title{Shift of biomass between juvenile and adult Size classes or 'growth function'
Valid only for Juvenile and adult}
\usage{
shift_next_sizeClass(L_mean, temperature, sex_params, size_width = 1)
}
\arguments{
\item{L_mean}{size mean of the max and min of the size class range \link{cm}}

\item{temperature}{\link{°C}}

\item{size_width}{defines how big are the size ranges \link{cm} e.g. 1: then each cm one size class.}
}
\value{
\link{day^-1}
}
\description{
Shift of biomass between juvenile and adult Size classes or 'growth function'
Valid only for Juvenile and adult
}
\examples{
shift_next_sizeClass(6.1, 10, parameters_solv$Fe_params)
}
