% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{shiftTo_juvenile}
\alias{shiftTo_juvenile}
\title{Shift of biomass from larvae to juvenile or 'growth function for larvae'}
\usage{
shiftTo_juvenile(Te)
}
\arguments{
\item{Te}{\link{cm}}
}
\value{
rate at which, larves change life stage to juvenile \link{day^-1}
}
\description{
Shift of biomass from larvae to juvenile or 'growth function for larvae'
}
\examples{
shiftTo_juvenile(10)
}
