% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClass.R
\name{solver_sizeClass_extended}
\alias{solver_sizeClass_extended}
\title{Solver for a size class model with Eggs, Larvae, 5 Juvenile classes and 2 Adult classes}
\usage{
solver_sizeClass_extended(t, state, parameters, temperature_dataSet)
}
\arguments{
\item{temperature_dataSet}{}
}
\value{
a df with solution to the equation system
}
\description{
Solver for a size class model with Eggs, Larvae, 5 Juvenile classes and 2 Adult classes
}
