% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_temming.R
\name{som_growth}
\alias{som_growth}
\title{Somatic growth for juvenile and adult}
\usage{
som_growth(b_length, temperature, max_size, sex)
}
\arguments{
\item{b_length}{current length \link{mm}}

\item{temperature}{current day temperature \link{°C}}

\item{max_size}{asymptotic size \link{mm}}

\item{sex}{distinct param. values for male and females 'F' or 'M'.}
}
\value{
new length
}
\description{
Somatic growth for juvenile and adult
}
\examples{
sg_juv(15, 20, 85, 'F')
}
