% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{som_growth_thesis}
\alias{som_growth_thesis}
\title{Somatic growth developed in this Thesis as function of time}
\usage{
som_growth_thesis(L0, time, temp, sex)
}
\arguments{
\item{L0}{initial or current size \link{cm}}

\item{time}{time step}

\item{sex}{'F' or 'M'}

\item{Tem}{temperature \link{°C}}
}
\value{
new size after gorwth dependent on temperature
}
\description{
Somatic growth developed in this Thesis as function of time
}
\examples{
som_growth_thesis(3, 1, 5,"F") or:
development.df_juvI <- data.frame(row = seq(1, 1095))
simulate development with constant temperature for each 0.1°C & store results columns wise in the df:
for (i in temperature_range){
  development <- c()
  initial_l <- 0.6 #cm
  for (j in time_range){
    growth = som_growth_thesis(initial_l, j ,i , "F")
    development = append(development, growth) }
development.df_juvI[as.character(i)] <- development }

}
