% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{som_growthAF}
\alias{som_growthAF}
\title{Somatic growth in one time step of 1 day with K_func as TPC}
\usage{
som_growthAF(L, Tem, sex)
}
\arguments{
\item{L}{current size \link{cm}}

\item{Tem}{temperature \link{°C}}

\item{sex}{'F' or 'M'}
}
\value{
new size after gorwth dependent on temperature
}
\description{
Somatic growth in one time step of 1 day with K_func as TPC
}
\examples{
som_growthAF(3, 10, "F")
}
