% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClass.R
\name{spawning_rate}
\alias{spawning_rate}
\title{Rate of Biomass from Females that flows into egg biomass}
\usage{
spawning_rate(L, temperature)
}
\arguments{
\item{L}{size \link{cm}}

\item{T}{Temperature \link{°C}}
}
\description{
Rate of Biomass from Females that flows into egg biomass
}
