% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClass.R
\name{spawning_rate_outdated}
\alias{spawning_rate_outdated}
\title{*Deprecated Rate of Biomass from Females that flows into egg biomass}
\usage{
spawning_rate_outdated(L, temperature)
}
\arguments{
\item{L}{size \link{cm}}

\item{T}{Temperature \link{°C}}
}
\description{
*Deprecated Rate of Biomass from Females that flows into egg biomass
}
