% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{ST_predation}
\alias{ST_predation}
\title{Spatial-Temporal mortality model}
\usage{
ST_predation(t_day, Te, B_s, eta, beta = 9, sigma = 0.8)
}
\arguments{
\item{t_day}{simulation day}

\item{Te}{Temperature \link{°C}}

\item{B_s}{Biomass of the prey (shrimp)}

\item{eta}{eta function}

\item{sigma}{}
}
\description{
Spatial-Temporal mortality model
}
