% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{temperature_func}
\alias{temperature_func}
\title{Create specific data set for temperature within a time range}
\usage{
temperature_func(dataSet, startDate, endDate)
}
\arguments{
\item{dataSet}{from read.csv()}

\item{startDate}{beginning of wished time range as Character}

\item{endDate}{end of wished time range as Character}
}
\value{
temperature data set of the wished time period with two columns: 'date_time' and 'temperature'
}
\description{
Create specific data set for temperature within a time range
}
\examples{
temperature_func(dataSet ,"01/01/2000", "31/12/2000")
}
