% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{temperature_funcSolver}
\alias{temperature_funcSolver}
\title{Temperature funciton to be used within a solver}
\usage{
temperature_funcSolver(df, t)
}
\arguments{
\item{df}{dataFrame with temperature data of a limited time frame. It must content column 'temperature'}

\item{t}{time step from solver}
}
\value{
Temperature \link{°C}
}
\description{
Temperature funciton to be used within a solver
}
