% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{K_func_briere}
\alias{K_func_briere}
\title{K func based of flexTPC NB: for the time being all params only for fems}
\usage{
K_func_briere(temperature)
}
\arguments{
\item{temperature}{}
}
\value{
K value as function of temperature
}
\description{
K func based of flexTPC NB: for the time being all params only for fems
}
