% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{respiration_rate_b}
\alias{respiration_rate_b}
\title{Respiration rate}
\usage{
respiration_rate_b(temperature, L, sex_params)
}
\arguments{
\item{temperature}{Temperature \link{°C}}

\item{L}{size \link{cm}}
}
\value{
\link{gr / time }
}
\description{
Respiration rate
}
