% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{sel_probit}
\alias{sel_probit}
\title{Selectivity probability (Fishery)
#Rename it to Sel Logit !!! it has been changed}
\usage{
sel_probit(L, L50 = 4.49, SR = 1.56)
}
\arguments{
\item{SR}{}
}
\value{
the probability that a size class L will be catched by net
of characteristics L50 and SR
}
\description{
Selectivity probability (Fishery)
#Rename it to Sel Logit !!! it has been changed
}
\examples{
sel_probit(6) or sel_probit(L50 = 4.49 , SR = 1.56)
}
