% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticGrowth_sizeClasses_II.R
\name{spawning_rate_b}
\alias{spawning_rate_b}
\title{Rate of Biomass from Females that flows into egg biomass}
\usage{
spawning_rate_b(L)
}
\arguments{
\item{L}{size \link{cm}}

\item{T}{Temperature \link{°C}}
}
\description{
Rate of Biomass from Females that flows into egg biomass
}
